Contig viewer refactor.
Use cursor pagination instead of conventional pagiation.
Some small improvements.

Bugfix: Maps for studies with lots of sampels (400 request after page > 20)
