+++ src/js/modules/analysis/helpers.js
@@ -1,0 +1,45 @@
+const Commons = require('../../commons');
+const TAXONOMY_COLOURS = Commons.TAXONOMY_COLOURS;
+
+/**
+ * Create a display of the series color
+ * @param {number} i index of series color
+ * @return {string} display element
+ */
+export function getColourSquareIcon(i) {
+    const taxColor = Math.min(TAXONOMY_COLOURS.length - 1, i);
+    return '<div class=\'puce-square-legend\' style=\'background-color: ' +
+        // eslint-disable-next-line security/detect-object-injection
+        Commons.TAXONOMY_COLOURS[taxColor] + '\'></div>';
+}
+
+/**
+ * Converts series index to capped index, used to handle events
+ * @param {number} index
+ * @param {number} numSeries
+ * @return {number}
+ */
+export function getSeriesIndex(index, numSeries) {
+    if (index >= numSeries - 1) {
+        index = numSeries - 1;
+    }
+    return index;
+}
+
+/**
+ * Enable toggling of series visibility, sync'd across table of series and chart
+ * @param {jQuery.element} elem table row
+ * @param {HighCharts.series} series in chart
+ * @param {integer} index index of serie in chart
+ * @param {integer} numSeries total number of series
+ * @param {boolean} defaultVisibility true if series should be visible by default
+ */
+export function setTableRowAndChartHiding(elem, series, index, numSeries, defaultVisibility) {
+    series.setVisible(defaultVisibility);
+    if (index === numSeries - 1) {
+        ($(elem).parent().children().slice(numSeries - 1)).toggleClass(
+            'disabled-clickable');
+    } else {
+        $(elem).toggleClass('disabled-clickable');
+    }
+}
