+++ webpack.dev.config.js
@@ -38,10 +38,13 @@ module.exports = {
     devServer: {
         port: 9000,
         proxy: {
-            '/metagenomics/api/latest': {
-                target: 'https://www.ebi.ac.uk/',
+            '/metagenomics/api': {
+                target: 'http://localhost:8000/',
                 secure: false,
                 changeOrigin: true,
+                pathRewrite: {
+                    '^/metagenomics/api/latest': '/v1',
+                },
                 headers: {
                     "Access-Control-Allow-Origin": "*",
                     "Access-Control-Allow-Methods": "GET, POST, PUT, DELETE, PATCH, OPTIONS",
@@ -58,6 +61,12 @@ module.exports = {
                 target: 'http://localhost:9000/',
                 pathRewrite: {
                     '^/metagenomics/samples/.*$' : '/metagenomics/sample.html',
+                },
+            },
+            '/metagenomics/analyses': {
+                target: 'http://localhost:9000/',
+                pathRewrite: {
+                    '^/metagenomics/analyses/.*$' : '/metagenomics/analysis.html',
                 },
             },
         }
